pyproject.toml
[tool.poetry]
name = "pythonproject3"
version = "0.1.0"
description = "Лера сама писала 100%"
authors = ["Valeria Kuzmina <l3ra131007@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.2"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
pandas-stubs = "^2.2.2.240603"
stubs = "^1.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest-cov = "^5.0.0"


[tool.poetry.group.lint.dependencies]
black = "^24.3.0"
isort = "^5.13.2"
mypy = "^1.9.0"
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'